Update index argument type of getSubDevicePath method in XDP device abstractions to match xrt::device and HAL API

+++ src/runtime_src/xdp/profile/device/hal_device/xdp_hal_device.cpp
@@ -94,7 +94,7 @@ int HalDevice::readTraceData(void* traceBuf, uint32_t traceBufSz, uint32_t numSa
   return xclReadTraceData(mHalDevice, traceBuf, traceBufSz, numSamples, ipBaseAddress, wordsPerSample);
 }
 
-std::string HalDevice::getSubDevicePath(std::string& subdev, uint64_t index)
+std::string HalDevice::getSubDevicePath(std::string& subdev, uint32_t index)
 {
   constexpr size_t maxSz = 256;
   char buffer[maxSz];
