+++ src/runtime_src/xdp/profile/device/xdp_base_device.h
@@ -45,7 +45,7 @@ public:
 
   virtual void* getRawDevice() = 0 ;
 
-  virtual std::string getSubDevicePath(std::string& subdev, uint64_t index) = 0;
+  virtual std::string getSubDevicePath(std::string& subdev, uint32_t index) = 0;
 };
 
 }
