+++ src/runtime_src/xdp/profile/device/xdp_xrt_device.h
@@ -47,7 +47,7 @@ public:
 
   virtual void* getRawDevice() { return mXrtDevice ; }
 
-  virtual std::string getSubDevicePath(std::string& subdev, uint64_t index);
+  virtual std::string getSubDevicePath(std::string& subdev, uint32_t index);
 };
 }
